refactor #148 : FunctioonManager refactoring
