+++ Outline/Outline/Source/OutlineApp.swift
@@ -5,8 +5,8 @@
 //  Created by Hyunjun Kim on 10/13/23.
 //
 
-import Firebase
 import CoreLocation
+import Firebase
 import SwiftUI
 
 @main
