+++ Outline/Outline/Source/View/Share/ShareMainView.swift
@@ -10,7 +10,6 @@ import SwiftUI
 
 struct ShareMainView: View {
     @StateObject var runningManager = RunningManager.shared
-    @ObservedObject var homeTabViewModel: HomeTabViewModel
     @StateObject private var viewModel = ShareViewModel()
     
     let runningData: ShareModel
