+++ Outline/Outline/Source/ViewModel/RecordDetailViewModel.swift
@@ -44,7 +44,7 @@ class RecordDetailViewModel: ObservableObject {
             }
             
             runningData["킬로미터"] = String(format: "%.1f", healthData.totalRunningDistance/1000)
-            runningData["시간"] = formatDuration(healthData.totalTime)
+            runningData["시간"] = healthData.totalTime.formatDuration()
             runningData["평균 페이스"] = healthData.averagePace.formattedAveragePace()
             runningData["BPM"] = "\(Int(healthData.averageHeartRate))"
             runningData["칼로리"] = "\(healthData.totalEnergy)"
