--- Outline/Shared/AppDelegate.swift
@@ -1,18 +1,0 @@
-//
-//  AppDelegate.swift
-//  Outline
-//
-//  Created by 김하은 on 10/26/23.
-//
-
-import SwiftUI
-
-struct AppDelegate: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    AppDelegate()
-}
